[feature/storage] Move driver.php
